+++ apps/backend/src/models/domain.rs
@@ -53,7 +53,7 @@ pub struct Market {
     pub quote_ticker: String,
     pub tick_size: u128,    // Minimum price increment in quote atoms
     pub lot_size: u128,     // Minimum size increment in base atoms
-    pub min_size: u64,      // Minimum order size in base atoms
+    pub min_size: u128,     // Minimum order size in base atoms
     pub maker_fee_bps: i32, // Maker fee in basis points (0-10000)
     pub taker_fee_bps: i32, // Taker fee in basis points (0-10000)
 }
